refactoring due to the review
